requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-retry==2.9.1
aiosignal==1.4.0
alembic==1.17.2
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.75.0
anyio==4.12.0
asyncer==0.0.8
attrs==25.4.0
backoff==2.2.1
blinker==1.9.0
boto3==1.42.16
botocore==1.42.16
brotli==1.2.0
cachetools==6.2.4
certifi==2025.11.12
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
coloredlogs==15.0.1
colorlog==6.10.1
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
databricks-agents==1.9.0
databricks-ai-bridge==0.11.0
databricks-langchain==0.12.1
databricks-mcp==0.5.1
databricks-sdk==0.76.0
dataclasses-json==0.6.7
datasets==4.4.2
ddgs==9.10.0
deprecation==2.1.0
dill==0.4.0
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
docker==7.1.0
docstring-parser==0.17.0
dspy==2.6.27
dydantic==0.0.8
et-xmlfile==2.0.0
fake-useragent==2.2.0
fastapi==0.127.1
filelock==3.20.1
flashrank==0.2.10
flask==3.1.2
flask-cors==6.0.2
flatbuffers==25.12.19
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2025.10.0
gepa==0.0.22
gitdb==4.0.12
gitpython==3.1.45
google-auth==2.45.0
googleapis-common-protos==1.72.0
grandalf==0.8
graphene==3.4.3
graphql-core==3.2.7
graphql-relay==3.2.0
grpcio==1.76.0
grpcio-status==1.76.0
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huey==2.5.5
huggingface-hub==1.2.3
humanfriendly==10.0
hyperframe==6.1.0
identify==2.6.15
idna==3.11
importlib-metadata==8.7.1
iniconfig==2.3.0
itsdangerous==2.2.0
jinja2==3.1.6
jiter==0.12.0
jmespath==1.0.1
joblib==1.5.3
json-repair==0.54.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
langchain==1.2.0
langchain-anthropic==1.3.0
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.5
langchain-mcp-adapters==0.2.1
langchain-openai==1.1.6
langchain-tavily==0.2.15
langchain-text-splitters==1.1.0
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-checkpoint-postgres==3.0.2
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.1
langmem==0.0.30
langsmith==0.5.1
librt==0.7.5
litellm==1.75.9
loguru==0.7.3
lxml==6.0.2
magicattr==0.1.6
mako==1.3.10
markdown-it-py==4.0.0
markupsafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.8
mcp==1.25.0
mdurl==0.1.2
mlflow==3.8.1
mlflow-skinny==3.8.1
mlflow-tracing==3.8.1
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
mypy==1.19.1
mypy-extensions==1.1.0
nest-asyncio==1.6.0
nodeenv==1.10.0
numpy==1.26.4
onnxruntime==1.23.2
openai==2.14.0
openevals==0.1.3
openpyxl==3.1.5
opentelemetry-api==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
optuna==4.6.0
orjson==3.11.5
ormsgpack==1.12.1
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.5.1
pluggy==1.6.0
pre-commit==4.5.1
primp==0.15.0
propcache==0.4.1
protobuf==6.33.2
psycopg==3.3.2
psycopg-binary==3.3.2
psycopg-pool==3.3.0
py4j==0.10.9.7
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.12.0
pygments==2.19.2
pyjwt==2.10.1
pyparsing==3.3.1
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.21
pytz==2025.2
pyyaml==6.0.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rpds-py==0.30.0
rsa==4.9.1
ruff==0.14.10
s3transfer==0.16.0
scikit-learn==1.8.0
scipy==1.16.3
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
socksio==1.0.0
sqlalchemy==2.0.45
sqlparse==0.5.5
sse-starlette==3.0.4
starlette==0.50.0
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
tomli==2.3.0
tqdm==4.67.1
trustcall==0.0.39
typer-slim==0.21.0
typing-extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.2
tzdata==2025.3
ujson==5.11.0
unitycatalog-ai==0.3.2
unitycatalog-client==0.3.1
unitycatalog-langchain==0.3.0
urllib3==2.6.2
uuid-utils==0.12.0
uvicorn==0.40.0
virtualenv==20.35.4
werkzeug==3.1.4
whenever==0.7.3
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
